fix(ci): line-limit guard — split oversized modules below cap
